make it better

+++ server.js
@@ -4,7 +4,7 @@ const bodyParser = require('body-parser');
 const mongoose = require('mongoose');
 const jwt = require('jsonwebtoken');
 const connectionString = "mongodb+srv://jiratchbelt:jb030339@animeapp-w2qbn.gcp.mongodb.net/AnimeApp?retryWrites=true&w=majority";
-
+const number = 1;
 var Token;
 const port = process.env.PORT || 3000;
 
@@ -113,49 +113,58 @@ app.get('/', (req, res) => {
 
 app.get('/AnimeData', (req, res) => {
 
-    Anime_model.find((err, data) => {
+    if(number==1){
+        
+        Anime_model.find((err, data) => {
         if (err) { console.log("Error getAnime"); throw err; }
         res.json(data)
+        console.log("Animedata")
+        console.log("BigAnimedata")
     })
+}
 
 });
 
 app.get('/api/AnimeApp/Users', (req, res) => {
 
-    Users_model.find((err, data) => {
+    if(number==1){ Users_model.find((err, data) => {
         if (err) { console.log("Error getdata"); throw err; }
         res.json(data)
-    })
+        console.log("User")
+    })}
 });
 
 app.get('/api/AnimeApp/GetReviews', (req, res) => {
 
-    Reviews_model.find((err, data) => {
+    if(number==1)  {Reviews_model.find((err, data) => {
         if (err) { console.log("Error getreviews"); throw err; }
         res.json(data)
-    })
+        console.log("Reviews")
+    })}
 });
 
 app.get('/api/AnimeApp/GetFavorites', (req, res) => {
 
-    Favorites_model.find((err, data) => {
+    if(number==1){  Favorites_model.find((err, data) => {
         if (err) { console.log("Error getfav"); throw err; }
         res.json(data)
-    })
+        console.log("Favorites")
+    })}
 });
 
 //Method POST
 app.post('/addAnime', (req, res) => {
   
 
-    Anime_model.create(req.body, (err, doc) => {
+    if(number==1){ Anime_model.create(req.body, (err, doc) => {
         if (err) {
             console.log("Failed");
             throw err;
         }
         res.json(doc);
-       // console.log(doc);
-    });
+        console.log(doc);
+
+    });}
 });
 
 //
@@ -164,11 +173,12 @@ app.post('/api/AnimeApp/login', (req, res) => {
     var Password = req.body.Password;
 
 
-    Users_model.findOne({ Username: Username }, (err, user) => {
+    if(number==1)
+     { Users_model.findOne({ Username: Username }, (err, user) => {
         if (err) { console.log("Error "); throw err; }
 
         else if (!user) {
-            //  console.log("Username is not found");
+             console.log("Username is not found");
             res.json({ "Found": "Nouser" });
         }
         else {
@@ -176,13 +186,13 @@ app.post('/api/AnimeApp/login', (req, res) => {
          
             Users_model.findOne({ Username: Username, Password: Password }, (err, pass) => {
                 if (err) { console.log("Error "); throw err; }
-                else if (!pass) { res.json({ "Found": "wrongpass" }); }
+                else if (!pass) {  console.log("wrongpass"); res.json({ "Found": "wrongpass" }); }
                 else {
                     let privateKey="this is a private key";
                     jwt.sign({user: Username }, privateKey, function(err, jwttoken) {
                     Token=jwttoken; 
                     res.json({ "Found": "found","Token":Token,"Username":Username});
-            
+                    console.log("found");
                 
                  }); 
              
@@ -191,7 +201,7 @@ app.post('/api/AnimeApp/login', (req, res) => {
 
         }
 
-    })
+    })}
 
 })
 
@@ -202,10 +212,10 @@ app.post('/api/AnimeApp/signup', (req, res) => {
 
 
 
-    Users_model.findOne({ Username: Username }, (err, user) => {
+    if(number==1){ Users_model.findOne({ Username: Username }, (err, user) => {
         if (err) { console.log("Error "); throw err; }
         else if (user) {
-            //   console.log("Username is already used");
+               console.log("Username is already used");
             res.json({ "Use": "Used" });
         }
         else {
@@ -218,7 +228,7 @@ app.post('/api/AnimeApp/signup', (req, res) => {
             })
 
         }
-    });
+    });}
 });
 
 //POST addReviews
@@ -230,14 +240,14 @@ app.post('/api/AnimeApp/addReviews', (req, res) => {
     var Animetitle = req.body.Animetitle;
 
 
-    Reviews_model.create(req.body, (err, doc) => {
+    if(number==1){Reviews_model.create(req.body, (err, doc) => {
         if (err) {
             console.log("Failed");
             throw err;
         }
         res.json(doc);
-       // console.log(doc);
-    });
+        console.log(doc);
+    });}
 });
 //POST AddFavorite
 
@@ -245,10 +255,10 @@ app.post('/api/AnimeApp/addFavorite', (req, res) => {
     var Username = req.body.Username;
     var anime_title = req.body.anime_title;
 
-    Favorites_model.findOne({ Username: Username, anime_title:anime_title }, (err, data) => {
+    if(number==1){Favorites_model.findOne({ Username: Username, anime_title:anime_title }, (err, data) => {
         if (err) { console.log("Error "); throw err; }
         else if (data) {
-            //   console.log("Username is already used");
+               console.log("Username is already used");
             res.json({ "Add": "Added" });
         }
         else {
@@ -265,7 +275,7 @@ app.post('/api/AnimeApp/addFavorite', (req, res) => {
             })
 
         }
-    });
+    });}
 });
 //DELETE
 
@@ -274,7 +284,7 @@ app.post('/api/AnimeApp/deleteFavorite', (req, res) => {
     let anime_title = req.body.anime_title;
     let Username =  req.body.Username;
    
-    Favorites_model.deleteOne({ Username: Username, anime_title:anime_title},(err, data) => {
+    if(number==1){Favorites_model.deleteOne({ Username: Username, anime_title:anime_title},(err, data) => {
         if (err) { console.log("Error "); res.sendStatus(500);
                   throw err; }
         else{
@@ -282,7 +292,7 @@ app.post('/api/AnimeApp/deleteFavorite', (req, res) => {
             console.log(`delete doc ${anime_title} from user ${Username} successfully`);
         }
     
-    });
+    });}
 });
 
 app.post('/api/AnimeApp/deleteReviews', (req, res) => {
@@ -291,7 +301,7 @@ app.post('/api/AnimeApp/deleteReviews', (req, res) => {
     let Username =  req.body.Username;
     let Description = req.body.Description;
    
-    Reviews_model.deleteOne({ Username:Username,Animetitle:Animetitle,Description:Description},(err, data) => {
+    if(number==1){ Reviews_model.deleteOne({ Username:Username,Animetitle:Animetitle,Description:Description},(err, data) => {
         if (err) { console.log("Error "); res.sendStatus(500);
                   throw err; }
         else{
@@ -299,7 +309,7 @@ app.post('/api/AnimeApp/deleteReviews', (req, res) => {
             console.log(`delete reviews doc ${Animetitle} from user ${Username}  with Description = ${Description} successfully`);
         }
     
-    });
+    });}
 });
 
 //start server
